LayoutTests/platform/blackberry/fast/events/keydown-multiple-select-element-expected.txt

This tests that a <select> element can receive a DOM Keyboard Event and that such an event can change its selection state.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Tests that a <select> receives keydown events:
PASS Received keydown event.
Tests that a keydown events can change the selection in <select>:
PASS select2.options[select2.selectedIndex].value is "PASS"

TEST COMPLETE
